Removed unused parameters from the renderGameObjects funtion

+++ BasedCore/BEapp.cpp
@@ -71,7 +71,7 @@ void BEapp::run()
 
             std::vector<std::unique_ptr<BVKBuffer>> uboBuffers(BVKSwapchain::MAX_FRAMES_IN_FLIGHT);
 
-            for (int i = 0; i < uboBuffers.size(); i++)
+            for (size_t i = 0; i < uboBuffers.size(); i++)
             {
                 uboBuffers[i] = std::make_unique<BVKBuffer>(*appDevice.get(), sizeof(GlobalUbo), 1, VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT);
                 uboBuffers[i]->map();
